Reorganized folder structures
Reorganized code into folders since we were getting a lot of files.
