+++ src/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import {BrowserRouter as Router, Route} from 'react-router-dom';
-import Home from './pages/Home';
-import NavbarUser from './pages/NavbarUser';
-import NavbarAdmin from './pages/NavbarAdmin';
-import NavbarHome from './pages/NavbarHome';
-import Footer from './pages/Footer';
+import Home from './frontend/pages/Home';
+import NavbarUser from './frontend/components/app/NavbarUser';
+import NavbarAdmin from './frontend/components/app/NavbarAdmin';
+import NavbarHome from './frontend/components/app/NavbarHome';
+import Footer from './frontend/components/app/Footer';
 import './App.css';
 
 function App() {
