+++ src/backend/createDB/create_sample_data.js
@@ -1,4 +1,4 @@
-const con = require('../db')
+const con = require('../db.config')
 
 
 con.query("USE `acptheater`", function (err, result) {
