+++ src/backend/schema/CSC450 - Group 3 - ERD.jpeg
